Add an entry buffer customisation
Uses a header-line for the mode to show basic info about the entry, and
keeps the content area for the entry content itself.

Requires a feature of Elfeed not yet submitted upstream. However, the
package will work without that feature, just the customisation will not
be fully functional (the entry buffer will have a header line, and
mail-style headers too).

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>