Configure powerline only if it is loaded
Do not require powerline, but configure it opportunistically if it is
loaded.  The powerline package adds hooks on load, and does not offer
a way to disable it after loading, so it is better to leave it to the
user to load powerline or not.

* elfeed-goodies-search-mode.el (powerline): Delete requires.
(search-header/rhs, search-header/draw-wide, search-header/draw-tight)
(elfeed-goodies/search-header-draw): Wrap with (with-eval-after-load
'powerline ...).
* elfeed-goodies-show-mode.el (powerline): Delete requires.
(elfeed-goodies/entry-header-line): Wrap with (with-eval-after-load
'powerline ...).
(elfeed-goodies/show-mode-setup): Configure header-line-format only if
powerline is loaded.
* elfeed-goodies.el: Delete powerline from Package-Requires: in commentary.
(elfeed-goodies/setup): Configure elfeed-search-header-function only if
powerline is loaded.